Updated the file, added variable functionality
Made some changes. In line 36, the led should be on the left of the
bitshift, as the right-hand position is the number of bits to shift by.
Aside from that i added a function to convert the integer 'tim' from the
timerCount input variable to a hex value so we can set the comparison
flag value to whatever that integer is.